// src/generated/models/New_catalogitemsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const New_catalogitemsnew_category = {
  100000000: 'Office Supplies',
  100000001: 'IT Equipment',
  100000002: 'Access & Security',
  100000003: 'Stationery',
  100000004: 'Other'
} as const;
export type New_catalogitemsnew_category = keyof typeof New_catalogitemsnew_category;
export const New_catalogitemsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type New_catalogitemsstatecode = keyof typeof New_catalogitemsstatecode;
export const New_catalogitemsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type New_catalogitemsstatuscode = keyof typeof New_catalogitemsstatuscode;

export interface New_catalogitemsBase {
  importsequencenumber?: number;
  new_available?: boolean;
  new_catalogitemid: string;
  new_category?: New_catalogitemsnew_category;
  new_description?: string;
  new_itemname?: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: New_catalogitemsstatecode;
  statuscode?: New_catalogitemsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface New_catalogitems extends New_catalogitemsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  new_availablename?: string;
  new_categoryname?: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}